Add C++ end to end test for RPC and distributed autograd.
Adding an end to end test for running a simple training loop in C++
for the distributed RPC framework.

The goal of this change is to enable LeakSanitizer and potentially catch memory
leaks in the Future. Enabling LSAN with python multiprocessing is tricky and we
haven't found a solution for this. As a result, adding a C++ test that triggers
most of the critical codepaths would be good for now.

As an example, this unit test would've caught the memory leak fixed by:
https://github.com/pytorch/pytorch/pull/31030

Differential Revision: [D21112208](https://our.internmc.facebook.com/intern/diff/D21112208/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D21112208/)!

ghstack-source-id: 102447581
Pull Request resolved: https://github.com/pytorch/pytorch/pull/36893