In interpolate, only call _interp_output_size in one place
It looks like this was made a separate function because of the `dim` argument,
but that argument is always equal to `input.dim() - 2`.  Remove the argument
and consolidate all call sites into one.  This also means that this will be
called on paths that previously didn't call it, but all those cases throw
exceptions anyway.

Differential Revision: [D21209993](https://our.internmc.facebook.com/intern/diff/D21209993/)

ghstack-source-id: 102773596
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37168