Simplify precision-specification in tests.
Now that assertEquals considers dtypes in determining tolerance, most
tests don't need explicitly set precision.

Those that do are a few half precision tests on cuda. In this PR, those
are broken out to be handled explicitly, though we may also want to
consider further loosening the tolerance on half-precision.

ghstack-source-id: 2368ee115e0b4409dcc920f5466d7818018a948f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37181