add fmt
fmt is a formatting library for C++. It has several properties that make it nice
for inclusion in PyTorch:
- Widely used
- Basically copies how Python does it
- Support for all the compilers and platforms we care about
- Standards track (C++20)
- Small code size
- Header only
- The author works at FB (hi @vitaut).

This PR includes it as a submodule and sets up the build. It also
provides some example usage, including a `TORCH_CHECK_FMT` macro in c10
that should test all the builds we care about.

ghstack-source-id: b1af8ce3fab8d5af217ce3e69d93c7f2a91b97ac
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37241